+++ src/member/member.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
-import { ExcelService } from '../common/excel/excel.service';
 import { DatabaseModule } from '../database/database.module';
+import { ExcelService } from '../excel/excel.service';
 import { GroupModule } from '../group/group.module';
 import { MemberController } from './member.controller';
 import { memberProvider } from './member.provider';
